SQUASH: lpm-levels optimizations
cpuidle: lpm-levels: constify idle latency method

msm_cpuidle_get_deep_idle_latency is a method that simply returns 10.
There is no sense in using it in this way, so let's use a preprocessor
macro to keep it clean and tidy.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

cpuidle: lpm-levels: omit tracing and debugging

There's no need for the end-user to need (or want) to debug the lpm
driver. All this does is add overhead for no purpose.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

cpuidle: lpm-levels: abolish dt walkthrough

This commit removes all traces of the sysfs toggle to print the device
tree entries for lpm. This is useless for the user, and it's just
another unnecessary check that Qualcomm includes in their code.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

cpuidle: lpm-levels: omit menu_select sysfs param

Correct me if I'm wrong, but this sysfs parameter isn't used anywhere
(at least as of this tree). Since it's static, it compiles through
anyway, and the compiler doesn't attempt to omit it. Let's handle that
manually.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

cpuidle: lpm-levels: remove unused latency export

The lpm driver exports a relatively large function to determine the
latency of a CPU/cluster. Unfortunately, this exported function is never
utilized, and it requires more functions to be compiled. Let's omit this
function, as well as it's other useless dependancies.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

cpuidle: lpm-levels: refactor cpu count casts

Qualcomm, the true geniuses they are, decided to initially cast this
variable to an unsigned long, and then once again to an unsigned int for
each usage case. Not only does this allocate the bit width of an
unsigned long for an integer value, we also risk an alignment fault from
converting a long to an integer. Let's use the proper type for this,
uintprt_t.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

cpuidle: lpm-levels: omit unused TZ api

The method lpm_cpu_pre_pc_cb isn't ever mentioned besides it's
declaration and initialization. Let's remove this chunk of code, since
the compiler refuses to metion it as unused.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>

cpuidle: lpm-levels: fix redundant cpu map lock

The lpm probe function utilizes the get_online_cpus method at a time
where cpu-online status isn't significant. Let's put this call right
before it is required, in order to avoid locking the cpu map when
lpm_root_node is ERR_OR_NULL (which would call put_online_cpus).

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: YousefAlgadri <yusufgadrie@gmail.com>
Signed-off-by: Tiktodz <kuplemarkeple@gmail.com>